Clean up Channel and Executor JS bindings
This commit rearranges things a bit. We have a single Channel concept
with a sender and receiver that can be constructed from the JS side. In
addition, I introduce a JsSink concept, which can be constructed from JS
and implements the Sink trait on the Rust side. This allows Rust
functions to accept a Sink from JS and send the output of an arbitrary
stream to it. From the JS side we can handle the `send` and `close`
methods on the Sink trait with our own JS callbacks.

This commit also adjusts the executor to avoided repeated allocations of
a NotifyHandle Arc. We just allocate one Arc<NotifyHandle> per executor
now.